requirements.txt
tensorflow==2.15.0
opencv-python==4.8.0.76
numpy<2.0
streamlit==1.32.0
streamlit-webrtc==0.47.1
Pillow==10.0.0
